Arel::Nodes::CurrentTime
