+++ lib/arel/extensions.rb
@@ -1,4 +1,6 @@
 require 'arel/extensions/unknown'
+require 'arel/extensions/time_with_precision'
+require 'arel/extensions/current_time'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
