+++ lib/arel/extensions/current_time.rb
@@ -1,0 +1,22 @@
+# rubocop:disable Naming/MethodName
+# rubocop:disable Naming/UncommunicativeMethodParamName
+
+module Arel
+  module Nodes
+    class CurrentTime < TimeWithPrecision
+    end
+  end
+
+  module Visitors
+    class ToSql
+      def visit_Arel_Nodes_CurrentTime(o, collector)
+        collector << 'current_time'
+        collector << "(#{o.precision.to_i})" if o.precision
+        collector
+      end
+    end
+  end
+end
+
+# rubocop:enable Naming/MethodName
+# rubocop:enable Naming/UncommunicativeMethodParamName
