+++ lib/arel/extensions/time_with_precision.rb
@@ -1,0 +1,13 @@
+module Arel
+  module Nodes
+    class TimeWithPrecision < Arel::Nodes::Node
+      attr_reader :precision
+
+      def initialize(precision: nil)
+        super()
+
+        @precision = precision
+      end
+    end
+  end
+end
